refactor: move query-api folder to api for cleaner structure
- Move all files from src/query-api/ to src/api/
- Update import in connect.ts to use new path
- Remove old query-api directory
- Maintain all functionality with simplified path structure

The API module is now located at src/api/ for a cleaner and
more conventional directory structure.
